배포 스크립트 수정
