+++ scripts/PXL.js
@@ -1,34 +1,18 @@
-require('./App');
-
+const Enquirer = require('enquirer');
+const fs = require('fs');
 const input = JSON.parse(fs.readFileSync('build/contracts/PXL.json'));
-const contract = new web3.eth.Contract(input.abi, process.env.PXL_ADDRESS);
-const enquirer = new Enquirer();
+const contract = new web3.eth.Contract(input.abi);
+const replace = require('replace-in-file');
+const BigNumber = require('bignumber.js');
+const decimals = Math.pow(10, 18);
 
-if (process.env.PXL_ADDRESS) {
-    log(`PXL_ADDRESS : ${process.env.PXL_ADDRESS}`)
-} else {
-    error('PXL_ADDRESS : Not registered yet!')
-}
-
-const choices = process.env.PXL_ADDRESS ? ['deploy', 'mint', 'burn', 'transfer', 'addOwner'] : ['deploy'];
-const questions = [{
-    type: 'radio',
-    name: 'result',
-    message: 'Which function do you want to run?',
-    choices: choices
-}];
-enquirer.register('radio', require('prompt-radio'));
-enquirer.ask(questions)
-    .then((answers) => eval(answers.result)())
-    .catch((err) => log(err));
-
-const deploy = async () => {
+module.exports = async () => {
+    let enquirer = new Enquirer();
     enquirer.question('initialSupply', 'initialSupply');
     enquirer.question('confirmInitialSupply', 'initialSupply (confirm)');
     let answer = await enquirer.prompt(['initialSupply', 'confirmInitialSupply']);
     if (!Number(answer.initialSupply) || answer.initialSupply != answer.confirmInitialSupply) return;
 
-    let contract = new web3.eth.Contract(input.abi);
     contract.deploy({
         data: input.bytecode,
         arguments: [new BigNumber(answer.initialSupply * decimals)]
@@ -47,53 +31,3 @@ const deploy = async () => {
             })
         });
 };
-
-const mint = async () => {
-    enquirer.question('amount', 'amount');
-    enquirer.question('confirmAmount', 'amount (confirm)');
-    let answer = await enquirer.prompt(['amount', 'confirmAmount']);
-    if (!Number(answer.amount) || answer.amount != answer.confirmAmount) return;
-
-    contract.methods
-        .mint(new BigNumber(answer.amount * decimals))
-        .send(sendDefaultParams)
-        .then(receipt => log(receipt));
-}
-
-const burn = async () => {
-    enquirer.question('amount', 'amount');
-    enquirer.question('confirmAmount', 'amount (confirm)');
-    let answer = await enquirer.prompt(['amount', 'confirmAmount']);
-    if (!Number(answer.amount) || answer.amount != answer.confirmAmount) return;
-
-    contract.methods
-        .burn(new BigNumber(answer.amount * decimals))
-        .send(sendDefaultParams)
-        .then(receipt => log(receipt));
-}
-
-const transfer = async () => {
-    enquirer.question('to', 'address');
-    enquirer.question('amount', 'amount');
-    enquirer.question('confirmAmount', 'amount (confirm)');
-    let answer = await enquirer.prompt(['to', 'amount', 'confirmAmount']);
-    if (answer.to == '') return;
-    if (!Number(answer.amount) || answer.amount != answer.confirmAmount) return;
-
-    contract.methods
-        .transfer(answer.to, new BigNumber(answer.amount * decimals))
-        .send(sendDefaultParams)
-        .then(receipt => log(receipt));
-}
-
-const addOwner = async () => {
-    enquirer.question('address', 'address');
-    enquirer.question('confirmAddress', 'address (confirm)');
-    let answer = await enquirer.prompt(['address', 'confirmAddress']);
-    if (answer.address == '' || answer.address != answer.confirmAddress) return;
-
-    return contract.methods
-        .addOwner(answer.address)
-        .send(sendDefaultParams)
-        .then(receipt => log(receipt));
-}
