+++ scripts/TokenDistributor.js
@@ -1,26 +1,10 @@
-require('./App');
-
+const Enquirer = require('enquirer');
+const fs = require('fs');
 const input = JSON.parse(fs.readFileSync('build/contracts/TokenDistributor.json'));
-const enquirer = new Enquirer();
+const contract = new web3.eth.Contract(input.abi);
+const replace = require('replace-in-file');
 
-if (process.env.TOKEN_DISTRIBUTOR_ADDRESS) {
-    log(`TOKEN_DISTRIBUTOR_ADDRESS : ${process.env.TOKEN_DISTRIBUTOR_ADDRESS}`)
-} else {
-    error('TOKEN_DISTRIBUTOR_ADDRESS : Not registered yet!')
-}
-
-const questions = [{
-    type: 'radio',
-    name: 'result',
-    message: 'Which function do you want to run?',
-    choices: ['deploy']
-}];
-enquirer.register('radio', require('prompt-radio'));
-enquirer.ask(questions)
-    .then((answers) => eval(answers.result)())
-    .catch((err) => log(err));
-
-const deploy = async () => {
+module.exports = async () => {
     if (!process.env.PXL_ADDRESS) {
         error('PXL_ADDRESS is not registered. Please .env.{network} file update!')
         return;
@@ -31,7 +15,6 @@ const deploy = async () => {
     let answer = await enquirer.prompt(['confirmPXLAddress']);
     if (!answer.confirmPXLAddress) return;
 
-    let contract = new web3.eth.Contract(input.abi);
     contract.deploy({
         data: input.bytecode,
         arguments: [process.env.PXL_ADDRESS]
